Add files via upload
I used the fetch function to send a GET request to the API. If the response is successful, I converted it to JSON for easier handling. If there’s an error, it's logged to the console. then to process the data I made a loop through each record in the API response, create a table row for each one, and fill it with data. Each piece of data is sanitized to prevent security issues.
Searched a function called the sanitize function which ensures any data inserted into the page is safe by escaping harmful HTML or scripts, just incase it fails to work which happened to me.